Shuffle images on front-end not backend
The wrong image would be served first because of the way the
brief_listener works. I could change it to react to the socket for the
first image, if it were important to do this all on the backend. But
since knowing the order of the images shouldn't really give you any
advantage, you're probably fine with just doing it on the front-end